requirements.txt
beautifulsoup4
bs4
fastapi
jsonschema
jsonschema-specifications
langchain
langchainplus-sdk
matplotlib
numpy
openai
openapi-schema-pydantic
pandas
pydantic
python-dotenv
streamlit
typer
typing_extensions
typing-inspect
tiktoken
streamlit-js-eval
pymongo
discord
discord.py
aiohttp
aiosignal
PyPDF2